Added support for URL stream processing in Producer and created separate stream processing for Kafka and URL
